pyproject.toml
[tool.poetry]
name = "littlelemonapi-v2"
version = "0.1.0"
description = ""
authors = ["Terry Brooks, Jr. <Terry.Arthur@BrooksJr.com>"]
readme = "README.md"
packages = [{ include = "littlelemonapi_v2" }]


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
djangorestframework-csv = "^2.1.1"
djoser = "^2.2.0"
django-guardian = "^2.4.0"
django-debug-toolbar = "^4.1.0"
psycopg2-binary = "^2.9.6"
loguru = "^0.7.0"
django-admin-env-notice = "^1.0"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
django-dynamic-breadcrumbs = "^0.4.2"
drf-writable-nested = "^0.7.0"
furl = "^2.1.3"
crispy-bootstrap5 = "^2023.10"
django-localflavor = "^4.0"
django-credit-cards = "^0.4.1"
django-dotenv = "^1.4.2"



[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"
pylint = "^2.17.4"
black = "^23.3.0"
django-seed = "^0.3.1"
icecream = "^2.1.3"
pre-commit = "^3.3.2"
pytest = "^7.3.1"



[tool.poetry.group.test.dependencies]
django-webtest = "^1.9.11"
coverage = "^7.3.2"
django-discover-runner = "^1.0"
mock = "^5.1.0"
model-bakery = "^1.17.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
